requirements.txt
sqlalchemy
psycopg2-binary
win11toast
python-dotenv